server/requirements.txt
Werkzeug==2.2.2
Flask==2.1.0
gunicorn==20.1.0
ecdsa
boto3
pysha3==1.0b1
merkletools
google-cloud-storage
azure-storage-blob
azure-identity